+++ lib/expressions/dependency.rb
@@ -4,7 +4,7 @@ require 'expressions'
 
 module Low
   # A Dependency Expression requires LowType in order to be injected via a constructor.
-  # See "LowDependency[:dependency]" for traditional dependency injection without LowType.
+  # See "Dependencies[:dependency]" for traditional dependency injection without LowType.
   class Dependency < ::Expressions::Expression
     attr_reader :provider_key, :var_name
 
@@ -31,7 +31,7 @@ module Low
 
     def union_value(value)
       @provider_key = value
-      @var_name = LowDependency.var_name_via_namespace(value)
+      @var_name = Dependencies.var_name_via_namespace(value)
     end
   end
 end